kafka-mongodb/requirements.txt
pymongo==3.10.0
kafka-python